+++ Arduino/Zapatilla-IP/src/tomacorrientes.cpp
@@ -8,7 +8,7 @@ tomacorrientes::tomacorrientes(DATA &data, PINES &pin)
 
 void tomacorrientes::begin()
 {
-    for(int i=0 ;i<N ;i++) 
+    for(uint8_t i=0 ;i<N ;i++) 
     {
         pinMode(pin->tomas[i], OUTPUT);
         pinMode(pin->leds[i], OUTPUT);
@@ -16,12 +16,12 @@ void tomacorrientes::begin()
     }
 }
 
-void tomacorrientes::invertir(int p)
+void tomacorrientes::invertir(uint8_t p)
 {
     data->estTomas[p] = !data->estTomas[p];
     conm(p, data->estTomas[p]);
 }
-void tomacorrientes::conm(int p, bool estado)
+void tomacorrientes::conm(uint8_t p, bool estado)
 { 
     estado = !estado;
     digitalWrite(pin->tomas[p], estado);
